refactor: Extract SASL stream feature rendering into SaslStreamFeatures
Moves appendSASLFeatures, asSASLMechanisms and the two mechanism element builders out of SASLAuthentication, leaving it as the negotiation state machine. The three callers of appendSASLFeatures are updated; No behaviour change.

Moves the corresponding tests into SaslStreamFeaturesTest.
